+++ app.js
@@ -5,6 +5,8 @@ require("dotenv").config();
 
 const contactsRouter = require("./routes/api/contacts-routes");
 
+// const authRouter = require("./routes/api/auth-routes");
+
 const app = express();
 
 const formatsLogger = app.get("env") === "development" ? "dev" : "short";
@@ -13,6 +15,8 @@ app.use(logger(formatsLogger));
 app.use(cors());
 app.use(express.json());
 
+
+// app.use("/api/auth", authRouter);
 app.use("/api/contacts", contactsRouter);
 
 app.use((req, res) => {
