+++ helpers/HttpError.js
@@ -15,3 +15,10 @@ const errorMessages = {
 
 module.exports = HttpError;
 
+// const HttpError = (status, message = errorMessages[status]) => {
+//   const error = new Error(message);
+//   error.status = status;
+//   return error;
+// };
+
+// module.exports = HttpError;
