+++ models/contact.js
@@ -1,5 +1,5 @@
 const { Schema, model } = require("mongoose");
-// const Joi = require("joi");
+const Joi = require("joi");
 
 const { handleMongooseError } = require("../utils");
 
@@ -25,22 +25,22 @@ const contactSchema = new Schema(
 
 contactSchema.post("save", handleMongooseError);
 
-// const addSchema = Joi.object({
-//   name: Joi.string().required(),
-//   email: Joi.string().required(),
-//   phone: Joi.string().required(),
-//   favorite: Joi.boolean(),
-// });
+const addSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().required(),
+  phone: Joi.string().required(),
+  favorite: Joi.boolean(),
+});
 
-// const updateFavoriteSchema = Joi.object({
-//   favorite: Joi.boolean().required(),
-// });
+const updateFavoriteSchema = Joi.object({
+  favorite: Joi.boolean().required(),
+});
 
-// const schemas = {
-//   addSchema,
-//   updateFavoriteSchema,
-// };
+const schemas = {
+  addSchema,
+  updateFavoriteSchema,
+};
 
 const Contact = model("contact", contactSchema);
 
-module.exports = Contact;
+module.exports = {Contact, schemas};
