--- utils/contactValidationSchemas.js
@@ -1,20 +1,0 @@
-const Joi = require("joi");
-
-const addSchema = Joi.object({
-    name: Joi.string().required(),
-    email: Joi.string().required(),
-    phone: Joi.string().required(),
-    favorite: Joi.boolean(),
-  });
-  
-  const updateFavoriteSchema = Joi.object({
-    favorite: Joi.boolean().required(),
-  });
-
-  const schemas = {
-    addSchema,
-    updateFavoriteSchema,
-  };
-
-module.exports = schemas;
-//   module.exports = schemas;
